requirements.txt
dbt-duckdb>=1.9.4
faker>=37.0.0